fflag w/o -

+++ ft_fspec.c
@@ -3,7 +3,7 @@
 char    *ft_fspec(t_args *ag, t_flags *fl)
 {
     if (!fl->lnh)
-        return (ft_dtoa(fl, (float)va_arg(ag->args, double), fl->psn));
+        return (ft_dtoa(fl, va_arg(ag->args, double), fl->psn));
     else if (fl->lnh == 1)
         return (ft_dtoa(fl, va_arg(ag->args, double), fl->psn));
     else if (fl->lnh == 5)
